+++ app/components/files/file-share.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import ContactSelector from '../contacts/contact-selector';
-import { t } from '../utils/translator';
+import { t, tx } from '../utils/translator';
 import chatState from '../messaging/chat-state';
 import fileState from '../files/file-state';
 
@@ -11,7 +11,8 @@ export default class FileShare extends Component {
                 limit={1}
                 onExit={() => chatState.routerModal.discard()}
                 action={contacts => chatState.startChatAndShareFiles(contacts, fileState.currentFile)}
-                title={t('title_shareWith')} />
+                title={t('title_shareWith')}
+                inputPlaceholder={tx('title_TryUsernameOrEmail')} />
         );
     }
 }