+++ app/components/settings/display.js
@@ -18,11 +18,11 @@ const bgStyle = {
 };
 
 const spacer = {
-    height: 24
+    height: vars.spacing.medium.midi
 };
 
 const text = {
-    color: vars.txtMedium,
+    color: vars.txtLightGrey,
     marginBottom: vars.spacing.small.midi2x,
     marginLeft: vars.spacing.small.midi2x
 };
@@ -40,12 +40,15 @@ export default class Display extends SafeComponent {
                 <PreferenceToggleItem
                     reverse
                     property="limitInlineImageSize"
-                    title={tx('title_showLargeImages', { size: util.formatBytes(config.chat.inlineImageSizeLimit) })} />
+                    title={tx('title_showLargeImages', { size: util.formatBytes(config.chat.inlineImageSizeLimit) })}
+                    description={tx('title_imageTooBigCutoff', { size: util.formatBytes(config.chat.inlineImageSizeLimitCutoff) })} />
+                <View style={spacer} />
                 {<Text style={text}>{tx('title_urlPreview')}</Text>}
                 <WarningItem
                     content={tx('title_EnableUrlPreviewWarning')}
                     linkContent={tx('title_learnMore')}
                     link=""
+                    iconType="security"
                 />
                 <PreferenceToggleItem
                     property="externalContentEnabled"