+++ app/components/shared/avatar.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { observer } from 'mobx-react/native';
-import { View, Text, TouchableOpacity, Dimensions, LayoutAnimation } from 'react-native';
+import { View, Text, TouchableOpacity, Dimensions, LayoutAnimation, Linking } from 'react-native';
 import { observable, reaction } from 'mobx';
 import SafeComponent from '../shared/safe-component';
 import icons from '../helpers/icons';
@@ -17,6 +17,10 @@ import CorruptedMessage from './corrupted-message';
 import tagify from './tagify';
 import { User } from '../../lib/icebear';
 import { tx } from '../utils/translator';
+import preferenceStore from '../settings/preference-store';
+import { popupSetupVideo } from '../shared/popups';
+import CircleButtonWithIcon from '../controls/circle-button-with-icon';
+
 
 const itemStyle = {
     flex: 1,
@@ -70,6 +74,14 @@ const nameMessageContainerStyle = {
     paddingTop: 0
 };
 
+const videoCallMsgStyle = {
+    color: 'rgba(0,0,0,.38)'
+};
+
+const linkStyle = {
+    color: '#2F80ED'
+};
+
 const nameTextStyle = {
     color: vars.txtMedium
 };
@@ -105,6 +117,14 @@ const lastMessageTextStyle = {
 
 const systemMessageStyle = {
     fontStyle: 'italic'
+};
+
+const lastMessageWithIcon = {
+    fontWeight: vars.font.weight.regular,
+    color: vars.txtMedium,
+    fontSize: vars.font.size.normal,
+    lineHeight: 22,
+    borderWidth: 0
 };
 
 const { width } = Dimensions.get('window');
@@ -177,7 +197,35 @@ export default class Avatar extends SafeComponent {
     }
 
     get systemMessage() {
-        const { systemMessage } = this.props;
+        const { systemMessage, videoCallMessage } = this.props;
+        if (videoCallMessage) {
+            const { prefs } = preferenceStore;
+            const videoCallShort = videoCallMessage.replace(/(https:\/\/)/, '');
+            return (
+                <View>
+                    <View style={{ flex: 1, flexDirection: 'row' }}>
+                        <Text style={[lastMessageWithIcon, videoCallMsgStyle]}>
+                            {systemMessage}
+                        </Text>
+                        <CircleButtonWithIcon
+                            name="info"
+                            iconColor="gray"
+                            onPress={() => popupSetupVideo(videoCallMessage)}
+                            radius={vars.iconSize}
+                        />
+                    </View>
+                    <TouchableOpacity onPress={() => prefs.hasSeenJitsiSuggestionPopup ? Linking.openURL(videoCallMessage) : popupSetupVideo(videoCallMessage)}
+                        pressRetentionOffset={vars.pressRetentionOffset}>
+                        <View style={{ flex: 1, flexDirection: 'row' }}>
+                            {icons.plaindark('videocam', vars.iconSizeSmall)}
+                            <Text style={[linkStyle]}>
+                                {videoCallShort}
+                            </Text>
+                        </View>
+                    </TouchableOpacity>
+                </View>
+            );
+        }
         return systemMessage && (
             <Text style={[lastMessageTextStyle, systemMessageStyle]}>
                 {systemMessage}